seperated netty transport unit tests into seperate file, fixed missing test method annotations
